% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetupr-package.R
\docType{package}
\name{meetupxlanimate-package}
\alias{meetupxlanimate}
\alias{meetupxlanimate-package}
\title{meetupxlanimate: meetupxlanimate}
\description{
R package for Meetup group admins only: 
             creates an Excel workbook with group data and 
             animation of membership growth. This package 
             currently uses code borrowed from the rladies 
             meetupr package as an expedient.
}
\author{
\strong{Maintainer}: Rick Pack \email{rickeyhp@gmail.com}

}
\keyword{internal}
