% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetupxlanimate.R
\name{meetupxlanimate}
\alias{meetupxlanimate}
\title{Get Meetup members' join date and other data, create animation}
\usage{
meetupxlanimate(meetupgrp_name)
}
\value{
A tibble with the following columns:
\itemize{
\item id
\item name
\item description
\item created
\item updated
\item post_count
\item discussion_count
\item latest_reply_created
\item latest_reply_member_name
\item resource
}
}
\description{
Get Meetup members' join date and other data, create animation
}
\examples{
\dontrun{
meetupgrp_name <- "tripass"
join_dates     <- meetupxlanimate(meetupgrp_name)
}
}
