% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rExcelhelper.R
\docType{data}
\name{midStyle}
\alias{midStyle}
\title{openxlsx helpful functions for meetupxlanimate presentation}
\format{
An object of class \code{Style} of length 1.
}
\usage{
midStyle
}
\description{
openxlsx helpful functions for meetupxlanimate presentation
}
\keyword{datasets}
