% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{rta_data}
\alias{rta_data}
\title{Example data for Research Triangle Analysts and
Triangle SQL Server User Group (TriPASS)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2645 rows and 12 columns.
}
\usage{
rta_data
}
\description{
Example data for Research Triangle Analysts and
Triangle SQL Server User Group (TriPASS)
}
\keyword{datasets}
